feat: add service productions

+++ models/producer/producer.go
@@ -3,7 +3,7 @@ package producer
 import "time"
 
 type Producer struct {
-	ID             uint      `json:"id,omitempty"`
+	ID             int64     `json:"id,omitempty"`
 	FirstName      string    `json:"first_name,omitempty"`
 	LastName       string    `json:"last_name,omitempty"`
 	DocumentNumber string    `json:"document_number,omitempty"`
